requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==0.61.1
django-cors-headers==4.6.0
django-filter==24.3
django-multiselectfield==0.1.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.3.0
